Move commands folder under src
