+++ src/commands/maps/status/index.js
