+++ src/commands/maps/status/stop/index.js
